refactor: Split socketio in multiple logical files
